+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Change Log
 ==========
 
+0.0.2 - Alpha
+-------------
+* Feature: FileStore wrapper around provider files
+* Bugfix: S3 url generation
+
 0.0.1 - Alpha
 -------------
 * Feature: Local File Storage